backend/requirements.txt
Flask==2.1.0
mysql-connector-python==8.0.27
firebase-admin==5.2.0
flask-cors==5.0.0
